# src/handlers/swagger/openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Lockbot",
    "description": "Coordinate use of your team's shared resources, in Slack 🤝 ... or via HTTP API",
    "contact": {
      "name": "Lockbot.app",
      "url": "https://lockbot.app/"
    },
    "version": "1.0.0"
  },
  "externalDocs": {
    "description": "GitHub Repository",
    "url": "https://github.com/connorads/lockbot"
  },
  "paths": {
    "/teams/{team}/channels/{channel}/locks": {
      "get": {
        "security": [
          {
            "basicAuth": []
          }
        ],
        "tags": ["Locks"],
        "summary": "Fetch all locks 📜",
        "operationId": "fetchAllLocks",
        "parameters": [
          {
            "name": "team",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "channel",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of active locks",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/locks"
                }
              }
            }
          }
        }
      },
      "post": {
        "security": [
          {
            "basicAuth": []
          }
        ],
        "tags": ["Locks"],
        "summary": "Create lock 🔒",
        "operationId": "createLock",
        "parameters": [
          {
            "name": "team",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "channel",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/lock"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Lock exists",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/lock"
                }
              }
            }
          },
          "201": {
            "description": "Lock created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/lock"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/error"
                }
              }
            }
          },
          "403": {
            "description": "Cannot create lock",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/error"
                }
              }
            }
          }
        }
      }
    },
    "/teams/{team}/channels/{channel}/locks/{lock}": {
      "get": {
        "security": [
          {
            "basicAuth": []
          }
        ],
        "tags": ["Locks"],
        "summary": "Fetch lock 👀",
        "operationId": "fetchLock",
        "parameters": [
          {
            "name": "team",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "channel",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "lock",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Lock found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/lock"
                }
              }
            }
          },
          "404": {
            "description": "Lock not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/error"
                }
              }
            }
          }
        }
      },
      "delete": {
        "security": [
          {
            "basicAuth": []
          }
        ],
        "tags": ["Locks"],
        "summary": "Delete lock 🔓",
        "operationId": "deleteLock",
        "parameters": [
          {
            "name": "team",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "channel",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "lock",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Lock deleted or doesn't exist"
          },
          "403": {
            "description": "Cannot delete lock",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/error"
                }
              }
            }
          },
          "404": {
            "description": "Lock not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/error"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "lock": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "owner": {
            "type": "string"
          }
        },
        "example": { "name": "dev", "owner": "U012345MNOP" }
      },
      "locks": {
        "type": "array",
        "items": {
          "$ref": "#/components/schemas/lock"
        },
        "example": [
          { "name": "dev", "owner": "U012345MNOP" },
          { "name": "test", "owner": "U012345MNOP" }
        ]
      },
      "error": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string"
          }
        },
        "example": { "error": "Insert specific error message here" }
      }
    },
    "securitySchemes": {
      "basicAuth": {
        "type": "http",
        "scheme": "basic"
      }
    }
  }
}